backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langchain
langchain_community
openai
langchain-openai
firebase-admin
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.16.0
